requirements.txt
alpaca-py
supabase
python-dotenv
pandas
ta-lib